requirements.txt
flask
openai
python-dotenv
beautifulsoup4
requests
